Произвел модуляцию проекта
